requirements.txt
openai
whisper
sounddevice
scipy
python-dotenv
gtts
